==> 06_Test_Best_Model.ipynb <==
# Test Best Model
In this notebook, we recover into Python the best trained model and evaluate it on the test data.

The steps in this notebook are
- [import libraries](#import),
- [read in the Azure ML workspace](#workspace),
- [download the best trained model](#download), and
- [evaluate the result using the test data](#evaluate)

## Imports  <a id='import'></a>

import os
from azureml.core import Workspace
from azureml.core.model import Model
import azureml.core
from get_auth import get_auth
print('azureml.core.VERSION={}'.format(azureml.core.VERSION))

## Read in the Azure ML workspace  <a id='workspace'></a>
Read in the the workspace created in a previous notebook.

auth = get_auth()
ws = Workspace.from_config(auth=auth)
ws_details = ws.get_details()
print('Name:\t\t{}\nLocation:\t{}'
      .format(ws_details['name'],
              ws_details['location']))

## Download the best trained model  <a id='download'></a>
Get the best trained model.

model_name = "FAQ_ranker"
model = Model(ws, name=model_name)
print(model.name, model.version, sep = '\t')

Download the model.

model.download(target_dir="outputs", exist_ok=True)

## Evaluate the model using the test data <a id='evaluate'></a>
Look at the model's performance on the held-aside test data. This can take a couple of minutes.

%run -t scripts/TestClassifier.py --model $model_name

This completes the hyperparameter tuning tutorial. You may be interested in the [MLAKSDeployAML tutorial](https://github.com/Microsoft/MLAKSDeployAML) for an example of deploying a trained Python model for scoring at scale.

The [next notebook](07_Train_With_AML_Pipeline.ipynb) shows how to create an Azure Machine Learning pipeline that can be called to do the work of notebooks [04_Hyperparameter_Random_Search](04_Hyperparameter_Random_Search.ipynb) and [05_Train_Best_Model](05_Train_Best_Model.ipynb) of this tutorial.
