==> 07_Tear_Down.ipynb <==
# Tear Down

In this notebook we delete the resources created for this tutorial.

The steps are
- [import the libraries](#import),
- [read in the Azure ML workspace](#workspace), and
- [delete the resources](#delete).

## Imports <a id='import'></a>

import os
from azureml.core import Workspace
import azureml.core
from get_auth import get_auth
print('azureml.core.VERSION={}'.format(azureml.core.VERSION))

## Read in the Azure ML workspace <a id='workspace'></a>
Read in the the workspace created in a previous notebook.

auth = get_auth()
ws = Workspace.from_config(auth=auth)
ws_details = ws.get_details()
print('Name:\t\t{}\nLocation:\t{}'
      .format(ws_details['name'],
              ws_details['location']))

Pull out the name of the resource group.

resource_group = ws.resource_group
resource_group

## Delete resources <a id='delete'></a>

Delete all the compute targets in an experiment.

for name, compute_target in ws.compute_targets.items():
    print('Deleting {} {}'.format(type(compute_target), name))
    compute_target.delete()

Delete the workspace. This will also delete the experiments and runs inside it.

ws.delete()

Delete the resource group. This will delete all the resources inside it, including those listed above. This step may take multiple minutes.

!az group delete --yes --name $resource_group

You have run through all the notebooks of the [MLHyperparameterTuning tutorial](https://github.com/Microsoft/MLHyperparameterTuning).
